refactor: 게시글 목록을 조회하는 경우 PostRepository 사용하도록 변경, PostLikeRepository는 찜만 관리
