+++ src/main/java/econo/buddybridge/post/repository/PostLikeRepositoryCustomImpl.java
@@ -1,17 +1,9 @@
 package econo.buddybridge.post.repository;
 
-import com.querydsl.core.types.OrderSpecifier;
-import com.querydsl.core.types.dsl.BooleanExpression;
 import com.querydsl.jpa.impl.JPAQueryFactory;
-import econo.buddybridge.matching.repository.MatchingRepository;
-import econo.buddybridge.post.dto.PostCustomPage;
-import econo.buddybridge.post.dto.PostResDto;
 import econo.buddybridge.post.entity.PostLike;
-import econo.buddybridge.post.entity.PostType;
-import econo.buddybridge.post.exception.PostInvalidSortValueException;
 import lombok.RequiredArgsConstructor;
 
-import java.util.List;
 import java.util.Optional;
 
 import static econo.buddybridge.post.entity.QPostLike.postLike;
@@ -20,7 +12,6 @@ import static econo.buddybridge.post.entity.QPostLike.postLike;
 public class PostLikeRepositoryCustomImpl implements PostLikeRepositoryCustom {
 
     private final JPAQueryFactory queryFactory;
-    private final MatchingRepository matchingRepository;
 
     @Override
     public Optional<PostLike> findByPostIdAndMemberId(Long postId, Long memberId) {
@@ -30,45 +21,5 @@ public class PostLikeRepositoryCustomImpl implements PostLikeRepositoryCustom {
                 .fetchOne();
 
         return Optional.ofNullable(like);
-    }
-
-    // Todo: 이거 PostRepositoryImpl로 보내버리기 -> PostLike는 Like만 관리
-    @Override
-    public PostCustomPage findPostsByLikes(Long memberId, Integer page, Integer size, String sort, PostType postType) {
-
-        List<PostResDto> content = queryFactory
-                .select(postLike.post)
-                .from(postLike)
-                .where(postLike.member.id.eq(memberId), buildPostTypeExpression(postType))
-                .offset((long) page * size)
-                .orderBy(buildOrderSpecifier(sort))
-                .fetch()
-                .stream()
-                .map(post -> new PostResDto(post, true, getMatchingDoneCount(post.getId())))
-                .toList();
-
-        Long totalElements = queryFactory
-                .select(postLike.count())
-                .from(postLike)
-                .where(postLike.member.id.eq(memberId))
-                .fetchOne();
-
-        return new PostCustomPage(content, totalElements, content.size() < size);
-    }
-
-    private Integer getMatchingDoneCount(Long postId) {
-        return matchingRepository.countMatchingDoneByPostId(postId);
-    }
-
-    private BooleanExpression buildPostTypeExpression(PostType postType) {
-        return postType != null ? postLike.post.postType.eq(postType) : null;
-    }
-
-    private OrderSpecifier<?> buildOrderSpecifier(String sort) {
-        return switch (sort.toLowerCase()) {
-            case "desc" -> postLike.post.createdAt.desc();
-            case "asc" -> postLike.post.createdAt.asc();
-            default -> throw PostInvalidSortValueException.EXCEPTION;
-        };
     }
 }
