+++ src/main/java/econo/buddybridge/post/service/PostLikeService.java
@@ -2,10 +2,8 @@ package econo.buddybridge.post.service;
 
 import econo.buddybridge.member.entity.Member;
 import econo.buddybridge.member.service.MemberService;
-import econo.buddybridge.post.dto.PostCustomPage;
 import econo.buddybridge.post.entity.Post;
 import econo.buddybridge.post.entity.PostLike;
-import econo.buddybridge.post.entity.PostType;
 import econo.buddybridge.post.repository.PostLikeRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -39,10 +37,5 @@ public class PostLikeService {
         PostLike newLike = new PostLike(post, member);
         postLikeRepository.save(newLike);
         return true;
-    }
-
-    @Transactional(readOnly = true)
-    public PostCustomPage getPostsLikes(Long memberId, Integer page, Integer size, String sort, PostType postType) {
-        return postLikeRepository.findPostsByLikes(memberId, page - 1, size, sort, postType);
     }
 }
